# alloul-one/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Alloul One API
  version: 0.1.0
  description: Core enterprise API contracts for multi-tenant platform.
servers:
  - url: https://api.alloulone.com/v1
security:
  - bearerAuth: []
paths:
...
  /orgs/{orgId}/knowledge/documents:
    get:
      summary: List knowledge documents
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '200': { description: Document list }
    post:
      summary: Create knowledge document
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, createdById]
              properties:
                title: { type: string }
                content: { type: object }
                createdById: { type: string }
      responses:
        '201': { description: Created }
  /orgs/{orgId}/search:
    get:
      summary: Search tenant data
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200': { description: Search results }
  /audit/logs:
    get:
      summary: List audit logs
      parameters:
        - in: query
          name: orgId
          schema: { type: string }
      responses:
        '200': { description: Audit logs }
    post:
      summary: Append audit log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, resourceType]
              properties:
                orgId: { type: string }
                actorUserId: { type: string }
                action: { type: string }
                resourceType: { type: string }
                resourceId: { type: string }
                payload: { type: object }
      responses:
        '201': { description: Logged }
  /audit/verify:
    get:
      summary: Verify immutable audit hash chain
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 500 }
      responses:
        '200': { description: Verification result }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    OrgId:
      in: path
      name: orgId
      required: true
      schema: { type: string }
...